Sort task by boards
