+++ data/database.json.dist
@@ -19,7 +19,7 @@
       "content": "git clone and npm install",
       "boardId": 1,
       "date": "2021-02-11",
-      "id": 2
+      "id": 1
     }
   ]
 }
