+++ src/pages/components/TaskGroups.jsx
@@ -3,15 +3,17 @@ import * as React from 'react'
 import { Box, Button, useDisclosure } from '@chakra-ui/react'
 // import { id } from 'date-fns/locale'
 
-import { getTaskGroups, createTask } from '../../utils/api'
+import { getTaskGroups, createTask, getTasks } from '../../utils/api'
 import Tasks from './Tasks'
 import { AddNewTask } from '../Board/AddNewTask'
+import { AddNewGroup } from '../Board/AddNewGroup'
 
 // eslint-disable-next-line react/prop-types
-export const TaskGroups = ({ boardId }) => {
+export const TaskGroups = ({ boardId, handleCreateGroup }) => {
   // eslint-disable-next-line no-unused-vars
   const [status, setStatus] = React.useState('loading')
   const [groups, setGroups] = React.useState([])
+  const [tasks, setTasks] = React.useState([])
   const [currentGroupId, setCurrentGroupId] = React.useState()
   const { isOpen, onOpen, onClose } = useDisclosure()
 
@@ -26,6 +28,18 @@ export const TaskGroups = ({ boardId }) => {
       console.log(e)
     }
   }, [boardId, groups])
+
+  React.useEffect(() => {
+    try {
+      const fetchData = async () => {
+        const data = await getTasks(boardId)
+        setTasks(data)
+      }
+      fetchData()
+    } catch (e) {
+      console.log(e)
+    }
+  }, [boardId, tasks])
 
   const handleCreateTasks = (newTask) => {
     createTask(boardId, currentGroupId, newTask)
@@ -58,7 +72,9 @@ export const TaskGroups = ({ boardId }) => {
             {group.name}
           </Box>
           <Box overflowY="scroll" h="300px">
-            <Tasks boardId={boardId} />
+            {tasks.map((task) => {
+              return group.taskIds.includes(task.id) ? <Tasks task={task} /> : null
+            })}
           </Box>
 
           <Box m="2">
@@ -80,6 +96,7 @@ export const TaskGroups = ({ boardId }) => {
           </Box>
         </Box>
       ))}
+      <AddNewGroup boardId={boardId} handleCreateGroup={handleCreateGroup} />
     </>
   )
 }
